update minecraft skin viewer
